refactor: modularize algorithm metadata
