+++ src/algorithms/sorting/index.ts
@@ -1,0 +1,5 @@
+import type { AlgoMeta } from "@/algorithms/types";
+import { bubbleSortMeta } from "./bubbleSort.meta";
+import { selectionSortMeta } from "./selectionSort.meta";
+
+export const sortingAlgos: AlgoMeta[] = [bubbleSortMeta, selectionSortMeta];
