+++ src/components/panels/CodePanel.tsx
@@ -10,7 +10,7 @@ import "highlight.js/styles/github.css";
 import "@/styles/hljs-dark-overrides.css";
 import "@/styles/codepanel.css";
 
-import type { AlgoMeta } from "@/engine/registry";
+import type { AlgoMeta } from "@/algorithms/types";
 
 hljs.registerLanguage("cpp", cpp);
 hljs.registerLanguage("java", java);
